[legacy] clean up amp
