+++ colossalai/legacy/amp/amp_type.py
