+++ colossalai/legacy/amp/torch_amp/_grad_scaler.py
